requirements.txt
Django==4.2.6

# DDBB
psycopg2==2.8.6 --no-binary psycopg2

# Environment
django-environ==0.9.0

# Rest
djangorestframework==3.14.0

# Celery
#tornado==6.2
redis==4.3
django-redis==5.2.0
celery==5.2.7
